+++ packages/components/src/CardList/CardList.test.tsx
@@ -4,11 +4,7 @@ import { describe, expect, it } from 'vitest'
 
 import { render } from '../test/utils'
 import { EMPTY_LIST, type IProps } from './CardList'
-import {
-  Default,
-  FiltedDisplay,
-  WhenFiltedDisplayIsZero,
-} from './CardList.stories'
+import { Default, WhenFiltedDisplayIsZero } from './CardList.stories'
 
 describe('CardList', () => {
   it('Shows all items when no filtering is done', () => {
@@ -17,15 +13,7 @@ describe('CardList', () => {
     expect(getAllByTestId('CardListItem').length).toBe(4)
   })
 
-  it('Shows only filted items when provided', () => {
-    const { getAllByTestId } = render(
-      <FiltedDisplay {...(FiltedDisplay.args as IProps)} />,
-    )
-
-    expect(getAllByTestId('CardListItem').length).toBe(2)
-  })
-
-  it('Shows the no items label when filted items is empty', () => {
+  it('Shows the no item label when filted items is empty', () => {
     const { getByText } = render(
       <WhenFiltedDisplayIsZero {...(WhenFiltedDisplayIsZero.args as IProps)} />,
     )
