+++ packages/components/src/CardList/CardList.tsx
@@ -3,13 +3,11 @@ import { Flex, Text } from 'theme-ui'
 
 import { CardListItem } from '../CardListItem/CardListItem'
 import { Icon } from '../Icon/Icon'
-import { Loader } from '../Loader/Loader'
 
 import type { IMapPin } from 'oa-shared'
 
 export interface IProps {
   columnsCountBreakPoints?: { [key: number]: number }
-  filteredList: IMapPin[] | null
   list: IMapPin[]
   onBlur: () => void
   onPinClick: (arg: IMapPin) => void
@@ -18,11 +16,10 @@ export interface IProps {
 }
 
 export const EMPTY_LIST = 'Oh nos! Nothing to show!'
+const DEFAULT_BREAKPOINTS = { 600: 1, 1100: 2, 1600: 3 }
 
 export const CardList = (props: IProps) => {
   const {
-    columnsCountBreakPoints,
-    filteredList,
     list,
     onBlur,
     onPinClick,
@@ -30,8 +27,8 @@ export const CardList = (props: IProps) => {
     viewport,
   } = props
 
-  const listToShow = filteredList === null ? list : filteredList
-  const displayItems = listToShow
+  const displayItems = list
+    .splice(0, 30)
     .sort(
       (a, b) =>
         Date.parse(b.creator?._lastActive || '0') -
@@ -50,11 +47,10 @@ export const CardList = (props: IProps) => {
       )
     })
 
-  const isListEmpty = displayItems.length === 0
-  const hasListLoaded = list
-  const results = `${displayItems.length} result${
-    displayItems.length == 1 ? '' : 's'
-  } in view`
+  const isListEmpty = list.length === 0
+  const results = `${list.length} result${list.length == 1 ? '' : 's'} in view`
+  const columnsCountBreakPoints =
+    props.columnsCountBreakPoints || DEFAULT_BREAKPOINTS
 
   return (
     <Flex
@@ -65,36 +61,25 @@ export const CardList = (props: IProps) => {
         padding: 2,
       }}
     >
-      {!hasListLoaded && <Loader />}
-      {hasListLoaded && (
-        <>
-          <Flex
-            sx={{
-              justifyContent: 'space-between',
-              paddingX: 2,
-              paddingTop: 2,
-              fontSize: 2,
-            }}
-          >
-            <Text data-cy="list-results">{results}</Text>
-            <Flex sx={{ alignItems: 'center', gap: 1 }}>
-              <Text> Most recently active</Text>
-              <Icon glyph="arrow-full-down" />
-            </Flex>
-          </Flex>
-          {isListEmpty && EMPTY_LIST}
-          {!isListEmpty && (
-            <ResponsiveMasonry
-              columnsCountBreakPoints={
-                columnsCountBreakPoints
-                  ? columnsCountBreakPoints
-                  : { 600: 1, 1100: 2, 1600: 3 }
-              }
-            >
-              <Masonry>{displayItems}</Masonry>
-            </ResponsiveMasonry>
-          )}
-        </>
+      <Flex
+        sx={{
+          justifyContent: 'space-between',
+          paddingX: 2,
+          paddingTop: 2,
+          fontSize: 2,
+        }}
+      >
+        <Text data-cy="list-results">{results}</Text>
+        <Flex sx={{ alignItems: 'center', gap: 1 }}>
+          <Text> Most recently active</Text>
+          <Icon glyph="arrow-full-down" />
+        </Flex>
+      </Flex>
+      {isListEmpty && EMPTY_LIST}
+      {!isListEmpty && (
+        <ResponsiveMasonry columnsCountBreakPoints={columnsCountBreakPoints}>
+          <Masonry>{displayItems}</Masonry>
+        </ResponsiveMasonry>
       )}
     </Flex>
   )
