--- src/pages/Maps/Content/index.ts
@@ -1,2 +1,0 @@
-export { MapView } from './MapView/MapView.client'
-export { Controls } from './Controls/Controls'
